Fix mongo support

+++ regolith/tools.py
@@ -1,14 +1,13 @@
 """Misc. regolith tools.
 """
 import os
-import re
 import sys
 import platform
 import email.utils
-from tinydb import Query
+import tinydb
 from datetime import datetime
 
-#import pymongo
+import pymongo
 
 if sys.version_info[0] >= 3:
     string_types = (str, bytes)
@@ -24,12 +23,13 @@ ON_MAC = (platform.system() == 'Darwin')
 ON_LINUX = (platform.system() == 'Linux')
 ON_POSIX = (os.name == 'posix')
 
-#if pymongo.version.split('.')[0] == '2':
-#    ON_PYMONGO_V2 = True
-#    ON_PYMONGO_V3 = False
-#else:
-ON_PYMONGO_V2 = False
-ON_PYMONGO_V3 = True
+if pymongo.version.split('.')[0] == '2':
+    ON_PYMONGO_V2 = True
+    ON_PYMONGO_V3 = False
+else:
+    ON_PYMONGO_V2 = False
+    ON_PYMONGO_V3 = True
+
 
 def fallback(cond, backup):
     """Decorator for returning the object if cond is true and a backup if cond is false.
@@ -37,6 +37,7 @@ def fallback(cond, backup):
     def dec(obj):
         return obj if cond else backup
     return dec
+
 
 @fallback(ON_PYMONGO_V2, None)
 class InsertOneProxy(object):
@@ -47,7 +48,8 @@ class InsertOneProxy(object):
 
 
 def insert_one(coll, doc):
-    return coll.insert(doc)
+    if isinstance(coll, tinydb.database.Table):
+        return coll.insert(doc)
 
     if ON_PYMONGO_V2:
         i = coll.insert(doc)
@@ -55,21 +57,27 @@ def insert_one(coll, doc):
     else:
         return coll.insert_one(doc)
 
+
 def insert_many(coll, docs):
-    return coll.insert_multiple(docs)
+    if isinstance(coll, tinydb.database.Table):
+        return coll.insert_multiple(docs)
 
     if ON_PYMONGO_V2:
         return coll.insert(docs)
     else:
         return coll.insert_many(docs)
 
+
 def delete_one(coll, doc):
-    return coll.remove(eids=[coll.get(Query()._id == doc["_id"]).eid])
+    if isinstance(coll, tinydb.database.Table):
+        eids = [coll.get(tinydb.Query()._id == doc["_id"]).eid]
+        return coll.remove(eids=eids)
 
     if ON_PYMONGO_V2:
         return coll.remove(doc, multi=False)
     else:
         return coll.delete_one(doc)
+
 
 def all_docs_from_collection_tinydb(client, collname):
     """Yield all entries in for all collections of a given name in a given database."""
@@ -81,7 +89,7 @@ def all_docs_from_collection_tinydb(client, collname):
             continue
         yield from db.table(collname).all()
 
-    
+
 def all_docs_from_collection_mongo(client, collname):
     """Yield all entries in for all collections of a given name in a given database."""
     for dbname in client.database_names():
@@ -106,7 +114,6 @@ def find_one_and_update(coll, filter, update, **kwargs):
         return coll.update(doc, update, **kwargs)
     else:
         return coll.find_one_and_update(filter, update, **kwargs)
-    
 
 
 MONTHS = {
@@ -147,10 +154,11 @@ MONTHS = {
     'dec': 12,
     'dec.': 12,
     'december': 12,
-    }
+}
 
-SHORT_MONTH_NAMES = (None, 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 
+SHORT_MONTH_NAMES = (None, 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul',
                      'Aug', 'Sept', 'Oct', 'Nov', 'Dec')
+
 
 def month_to_int(m):
     """Converts a month to an integer."""
@@ -166,7 +174,7 @@ def date_to_float(y, m, d=0):
     y = int(y)
     m = month_to_int(m)
     d = int(d)
-    return y + (m/100.0) + (d/100000.0)
+    return y + (m / 100.0) + (d / 100000.0)
 
 
 def date_to_rfc822(y, m, d=1):
